Biological Functional age/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
numpy>=1.24.0
